Split oversized test module; document and guard single-module design
Fixes the 5 failing build jobs on this PR. The v1.1.0 test additions pushed
test_focused_behaviour.py to 1056 lines, over the 1000-line C0302 cap:

  tests/unit/test_focused_behaviour.py:1:0: C0302: Too many lines in module

.pylintrc-tests deliberately does not disable C0302, unlike .pylintrc which
disables it for production with a justification - so test modules are meant to
stay under the cap. Adding a disable would defeat that; the module is split
instead.

The secret-detection tests move wholesale to tests/unit/test_secret_detection.py.
They are a coherent unit - exactly the v1.1.0 feature set - not an arbitrary cut
to satisfy a line count. Collected test count is unchanged at 454, confirming a
move rather than a rewrite. The now-unused ET and pytest imports are dropped
from test_focused_behaviour.py, restoring its original import-free header.

Note that pylint reports "rated at 10.00/10" while exiting 16 on C0302, so the
score line is not a pass signal; exit status is.

Also answers the question of whether redactor.py itself should be split: no.
The class is 65% of the file and its state is threaded throughout, so extracting
the easy parts leaves a 1654-line class and solves nothing. More importantly the
file runs standalone - copied alone to an empty directory, with no package
context, it redacts a config successfully. Relative imports would end that, and
for a tool trusted with secrets "read this one file, then run it" is worth more
than a tidier module tree.

That property was previously undocumented and unasserted. Now:

- The rationale moves from a .pylintrc comment into the module docstring,
  alongside a map of the file's seven sections.
- Section banners mark those seams, and the class docstring gains a method-group
  index noting that ordering inside redact_element() is load-bearing.
- tests/integration/test_standalone_script.py copies redactor.py into an empty
  directory and runs it there. Verified to fail on a simulated split: adding a
  relative sibling import fails all three tests, and removing it restores them.

456 tests pass, both pylint runs exit 0, reference snapshots untouched.
